backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
transformers==4.36.2
torch==2.1.0
spacy==3.7.2
google-generativeai==0.8.3
python-dotenv==1.0.1
